# examples/bookstore/generated/restapi.rs
// Code generated by oapi-codegen-rust. DO NOT EDIT.

#[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq)]
pub struct ListBooksQuery {
    /// Restrict the listing to a single author.
    #[serde(skip_serializing_if = "Option::is_none")]
    pub author: Option<String>,
    /// Only return books carrying every given tag.
    #[serde(skip_serializing_if = "Option::is_none")]
    pub tag: Option<Vec<String>>,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub limit: Option<i32>,
}

#[derive(Debug, Clone)]
pub struct CreateBookHeaders {
    /// Unique key so a retried create is not duplicated.
    pub idempotency_key: String,
}

impl<S> axum::extract::FromRequestParts<S> for CreateBookHeaders
where
    S: Send + Sync,
{
    type Rejection = (axum::http::StatusCode, String);
    async fn from_request_parts(
        parts: &mut axum::http::request::Parts,
        _state: &S,
    ) -> Result<Self, Self::Rejection> {
        let idempotency_key = match parts.headers.get("Idempotency-Key") {
            Some(value) => {
                let text = match value.to_str() {
                    Ok(text) => text,
                    Err(_) => {
                        return Err((
                            axum::http::StatusCode::BAD_REQUEST,
                            "header `Idempotency-Key` is not valid text".to_owned(),
                        ));
                    }
                };
                text.to_owned()
            }
            None => {
                return Err((
                    axum::http::StatusCode::BAD_REQUEST,
                    "missing required header `Idempotency-Key`".to_owned(),
                ));
            }
        };
        return Ok(Self { idempotency_key });
    }
}

/// Server behaviour: implement one method per operation.
pub trait Api: Clone + Send + Sync + 'static {
    /// List books, optionally filtered.
    fn list_books(
        &self,
        query: ListBooksQuery,
    ) -> impl std::future::Future<Output = ListBooksResponse> + Send;
    /// Add a book to the catalog.
    fn create_book(
        &self,
        headers: CreateBookHeaders,
        body: crate::apimodel::catalog::NewBook,
    ) -> impl std::future::Future<Output = CreateBookResponse> + Send;
    /// Fetch a single book by id.
    fn get_book(
        &self,
        id: String,
    ) -> impl std::future::Future<Output = GetBookResponse> + Send;
    /// Liveness probe returning a free-form document.
    fn get_health(&self) -> impl std::future::Future<Output = GetHealthResponse> + Send;
}

/// List books, optionally filtered.
pub enum ListBooksResponse {
    /// The matching books.
    Ok(Vec<crate::apimodel::catalog::Book>),
}

impl axum::response::IntoResponse for ListBooksResponse {
    fn into_response(self) -> axum::response::Response {
        match self {
            ListBooksResponse::Ok(body) => {
                const STATUS: axum::http::StatusCode = match axum::http::StatusCode::from_u16(
                    200,
                ) {
                    Ok(status) => status,
                    Err(_) => panic!("oapi-codegen emitted an invalid HTTP status code"),
                };
                (STATUS, axum::Json(body)).into_response()
            }
        }
    }
}

/// Add a book to the catalog.
pub enum CreateBookResponse {
    /// The book was created.
    Created(crate::apimodel::catalog::Book),
    /// The request was malformed.
    BadRequest(crate::apimodel::common::ErrorResponse),
}

impl axum::response::IntoResponse for CreateBookResponse {
    fn into_response(self) -> axum::response::Response {
        match self {
            CreateBookResponse::Created(body) => {
                const STATUS: axum::http::StatusCode = match axum::http::StatusCode::from_u16(
                    201,
                ) {
                    Ok(status) => status,
                    Err(_) => panic!("oapi-codegen emitted an invalid HTTP status code"),
                };
                (STATUS, axum::Json(body)).into_response()
            }
            CreateBookResponse::BadRequest(body) => {
                const STATUS: axum::http::StatusCode = match axum::http::StatusCode::from_u16(
                    400,
                ) {
                    Ok(status) => status,
                    Err(_) => panic!("oapi-codegen emitted an invalid HTTP status code"),
                };
                (STATUS, axum::Json(body)).into_response()
            }
        }
    }
}

/// Fetch a single book by id.
pub enum GetBookResponse {
    /// The requested book.
    Ok(crate::apimodel::catalog::Book),
    /// No resource matched the request.
    NotFound,
}

impl axum::response::IntoResponse for GetBookResponse {
    fn into_response(self) -> axum::response::Response {
        match self {
            GetBookResponse::Ok(body) => {
                const STATUS: axum::http::StatusCode = match axum::http::StatusCode::from_u16(
                    200,
                ) {
                    Ok(status) => status,
                    Err(_) => panic!("oapi-codegen emitted an invalid HTTP status code"),
                };
                (STATUS, axum::Json(body)).into_response()
            }
            GetBookResponse::NotFound => {
                const STATUS: axum::http::StatusCode = match axum::http::StatusCode::from_u16(
                    404,
                ) {
                    Ok(status) => status,
                    Err(_) => panic!("oapi-codegen emitted an invalid HTTP status code"),
                };
                STATUS.into_response()
            }
        }
    }
}

/// Liveness probe returning a free-form document.
pub enum GetHealthResponse {
    /// An opaque service-health document.
    Ok(serde_json::Value),
}

impl axum::response::IntoResponse for GetHealthResponse {
    fn into_response(self) -> axum::response::Response {
        match self {
            GetHealthResponse::Ok(body) => {
                const STATUS: axum::http::StatusCode = match axum::http::StatusCode::from_u16(
                    200,
                ) {
                    Ok(status) => status,
                    Err(_) => panic!("oapi-codegen emitted an invalid HTTP status code"),
                };
                (STATUS, axum::Json(body)).into_response()
            }
        }
    }
}

/// Build an axum `Router` that dispatches each route to `api`.
pub fn router<T: Api>(api: T) -> axum::Router {
    axum::Router::new()
        .route(
            "/books",
            axum::routing::get(list_books_handler::<T>).post(create_book_handler::<T>),
        )
        .route("/books/{id}", axum::routing::get(get_book_handler::<T>))
        .route("/health", axum::routing::get(get_health_handler::<T>))
        .with_state(api)
}

async fn list_books_handler<T: Api>(
    axum::extract::State(api): axum::extract::State<T>,
    axum_extra::extract::Query(query): axum_extra::extract::Query<ListBooksQuery>,
) -> ListBooksResponse {
    api.list_books(query).await
}

async fn create_book_handler<T: Api>(
    axum::extract::State(api): axum::extract::State<T>,
    headers: CreateBookHeaders,
    axum::Json(body): axum::Json<crate::apimodel::catalog::NewBook>,
) -> CreateBookResponse {
    api.create_book(headers, body).await
}

async fn get_book_handler<T: Api>(
    axum::extract::State(api): axum::extract::State<T>,
    axum::extract::Path(id): axum::extract::Path<String>,
) -> GetBookResponse {
    api.get_book(id).await
}

async fn get_health_handler<T: Api>(
    axum::extract::State(api): axum::extract::State<T>,
) -> GetHealthResponse {
    api.get_health().await
}